7월19일 오전11시50분 itemForm.html 상품등록 수정중 main.html 메인페이지 무한스크롤 적용됨
오후6시20분
main.html 수정

8월19일 오후2시30분
최신버전, discountService 추가

8월19일 저녁8시
ThymeleafExController, ItemDto, ItemFormDto,MainItemDto, OrderItemDto int에서 BigDecimal로 변경
	Item, Order, OrderItem int에서 BigDecimal로 변경
	할인율계산 메서드 DiscountService  추가
	itemForm.html 할인율입력박스 추가
	main.html 할인율 표시 수정중

8월20일
다시push

커밋 푸시완료

	ItemController   service에서 할인로직 가져와서 계산기능 추가
	MainItemDto discountrate, finalPrice 추가
	ItemRepositoryCustomImpl  discountrate, finalPrice 추가
	ItemService getItemById 추가
	main.html 할인율, 최종가격 추가
	최종가격 null로 표시되는문제 확인중

8월22일
	메인화면 상품 할인적용됬는데 소수점붙음, 전체상품은 윗1줄밖에 적용안됨 무한스크롤 ajax봐야할듯
	메인화면 상품 전부 할인적용, 소수점 정리 완료

	할인율로직 Service에서 Item엔티티로 변경 소수점,반올림 처리로직도 추가
	ItemRepositoryCustomImpl 레포지토리 수정
	DiscountService 엔티티로 로직옮겨서 안쓰는중
	main.html 베스트상품,전체상품, 무한스크롤 로직부분 할인율, 원래가격취소선, 할인된가격 추가,수정

	할인율로직 중복 피하고 재사용성을 좋게하기위해 엔티티에서 서비스로 옮길시도할 예정

8월23일
	Item엔티티에서 DiscountService로 할인율로직, 십의자리 반올림 로직 옮김
	현재 전체상품 무한스크롤부분부터 최종가격이 undefined원 으로 뜨는 오류 수정중

8월26일
	ItemRepositoryCustomImpl discountservice에서 할인로직 가져와 계산하고 dto에 설정하게 변경 -> 정상작동 확인

	상품상세정보
	ItemRepositoryCustomImpl discountservice에서 할인로직 가져와 계산하고 dto에 설정하게 변경 -> 정상작동 확인
	ItemController 최종가격 계산 추가
    	ItemFormDto 최종가격 필드 finalPrice 추가
    	ItemService getItemDtl 메서드 수정
    	ItemDtl.html 할인율,원래가격,할인가격 표시되고 적용되게 수정
	OrderService에 DiscountService 사용할수있게 주입
		order메서드에 최종가격 먼저계산후 주문아이템 생성시 최종가격 사용하게 수정,추가
	OrderItem의 createOrderItem메서드를 최종가격을 매개변수로 받아들일수 있게 수정,추가
		매개변수 finalPrice 추가, 주문가격을 최종가격으로 설정 추가
	OrderService의 orders 메서드도 최종가격 먼저계산후 주문아이템생성시 최종가격 사용하게 수정,추가

8월27일
   푸시
   CartDetailDto에서 itemId, finalPrice 추가
   	CartItemRepository에서 조회할 항목 상품id  i.id 추가
   	CartService에서 할인로직 계산,설정 추가
   	cartList.html 자바스크립트, html부분 price->finalPrice로 변경--
   	업데이트할때 BigDecimal 클래스 어떤역할이고 어떻게동작하고 db에는 어떻게들어가는지 설명하기

8월28일
   푸시

	MainController 데이터 전달 확인용 로그 수정
	Item엔티티 가격, 할인율 소수점 제거 시도한 코드 주석처리
	CartService BigDecimal로 인한 E표기법 문제해결을 위한 문자열 포맷팅 toPlainString 추가
	CartList.html toPlainString 추가

	mysql workbench에
    	ALTER TABLE item MODIFY price DECIMAL(10, 0) NOT NULL;
    	ALTER TABLE item MODIFY discountrate DECIMAL(10, 0);
    	로 가격, 할인율 테이블 타입 DECIMAL로 변경, 소수점자리 0으로 설정

8월29일
	푸시

	${cartItem.finalPrice.toPlainString() * cartItem.count}
	cartList 장바구니에 클릭하고 처음들어갈시 가격 상품개수만큼 붙어서나오는 오류 확인중

	템플릿에서 BigDecimal, String 둘다 곱셈연산 시 문제발생할수 있어 int타입으로 변경
    	cartItem.finalPrice를 int로 변환후 곱셈, 그 결과를 toString으로 변환하여 표시
    	확인완료

8월30일
    ItemFormDto 할인율 오류범위 0 미만 100 이상  추가

9월3일
   Bigdecimal 어차피 소수점 안쓸거라서 int로 전부수정. 표현식도 전부 수정

  Controller -
  ItemController, MainController,ThymeleafExController(실행오류 안나기위해변경)

  dto -
  CartDetailDto, ItemDto, ItemFormDto, MainItemDto, OrderItemDto

  entity -
  Item, Order, OrderItem

  repository -
  ItemRepositoryCustomImpl

  service -
  CartService, DiscountService, ItemService,
  OrderService

  templates -
  cartList.html, itemDtl.html, main.html

+++ src/main/java/com/shop/controller/ItemController.java
@@ -23,7 +23,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -138,8 +137,8 @@ public class ItemController {
         log.info("==================== Start:/item/{itemId} -> GetMapping ======================");
         ItemFormDto itemFormDto = itemService.getItemDtl(itemId);
 
-        // 최종 가격 계산 추가 8월26일
-        BigDecimal finalPrice = discountService.calculateFinalPrice(itemFormDto.createItem());
+        // 최종 가격 계산 추가 8월26일 int로 변경 9월3일
+        int finalPrice = discountService.calculateFinalPrice(itemFormDto.createItem());
         model.addAttribute("item",itemFormDto);
         log.info("==================== END:/item/{itemId} -> GetMapping ======================");
         return "item/itemDtl";
