+++ src/main/java/com/shop/dto/MainItemDto.java
@@ -4,7 +4,6 @@ import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,13 +12,13 @@ public class MainItemDto {
     private String itemNm;
     private String itemDetail;
     private String imgUrl;
-    private BigDecimal price; // int -> BigDecimal 변경 8월19일
-    private BigDecimal discountrate; // 8월20일 추가
+    private int price; // int -> BigDecimal 변경 8월19일 // int로 다시 수정 9월3일
+    private int discountrate; // 8월20일 추가 // int로 다시 수정 9월3일
     private int stockNumber;
-    private BigDecimal finalPrice; // 최종 가격 필드 추가 8월20일
+    private int finalPrice; // 최종 가격 필드 추가 8월20일 // int로 다시 수정 9월3일
 
     @QueryProjection //Querydsl 결과 조회 시 MainItemDto 객체로 바로 오도록  활용
-    public MainItemDto(Long id, String itemNm, String itemDetail, String imgUrl, BigDecimal price, BigDecimal discountrate, int stockNumber, BigDecimal finalPrice){ // int -> BigDecimal 변경 8월19일
+    public MainItemDto(Long id, String itemNm, String itemDetail, String imgUrl, int price, int discountrate, int stockNumber, int finalPrice){ // int -> BigDecimal 변경 8월19일
         this.id = id;
         this.itemNm = itemNm;
         this.itemDetail = itemDetail;
