+++ src/main/java/com/shop/dto/OrderItemDto.java
@@ -4,14 +4,13 @@ import com.shop.entity.OrderItem;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
 public class OrderItemDto {
     private String itemNm;
     private int count;
-    private BigDecimal orderPrice; // int -> BigDecimal 변경 8월19일
+    private int orderPrice; // int -> BigDecimal 변경 8월19일 // 다시int로변경 9월3일
     private String imgUrl;
     public OrderItemDto(OrderItem orderItem, String imgUrl){
         this.itemNm = orderItem.getItem().getItemNm();
